// src/constants/enums.ts
export enum StatusResponseEnum {
    FAILED = 'FAILED',
    SUCCESS = 'SUCCESS'
}

export enum MessageResponseEnum {
    USERNAME_IS_REQUIRED = 'USERNAME IS REQUIRED',
    PASSWORD_IS_REQUIRED = 'PASSWORD IS REQUIRED',
    USERNAME_ALREADY_EXISTS = 'USERNAME ALREADY TAKEN',
    USER_CREATED_SUCCESSFULLY = 'USER CREATED SUCCESSFULLY',
    INTERNAL_SERVER_ERROR = 'INTERNAL SERVER ERROR',
    USER_DOES_NOT_EXISTS = 'USER DOES NOT EXISTS',
    INAVLID_CREDENTIALS = "INVALID CREDENTIALS",
    AUTHENTICATION_SUCCESSFUL = 'AUTHENTICATION SUCCESSFUL',
    NEW_PASSWORD_IS_REQUIRED = 'NEW PASSWORD IS REQUIRED',
    PASSWORD_CANNOT_BE_SAME = 'PASSWORD CAN NOT BE SAME',
    PASSWORD_NOT_UPDATED = 'PASSWORD NOT UPDATED',
    PASSWORD_UPDATED_SUCCESSFULLY = 'PASSWORD UPDATED SUCCESSFULLY',
    UNAUTHORIZED_USER = 'UNAUTHORIZED USER',
    NO_ACCESS_TOKEN_PROVIDED = 'NO ACCESS TOKEN PROVIDED',
    ID_IS_NOT_PROVIDED = 'ID IS NOT PROVIDED',
    CATEGORY_IS_NOT_PROVIDED = 'CATEGORY IS NOT PROVIDED',
    CONTENT_IS_NOT_PROVIDED = 'CONTENT IS NOT PROVIDED',
    UPLOADED_SUCCESSFULLY = 'UPLOADED SUCCESSFULLY',
    AUTHORIZED_USER = "AUTHORIZED USER",
    DIFFERENT_USER = 'DIFFERENT USER',
    UNABLE_TO_POST = 'UNABLE TO POST',
    NEW_USERNAME_IS_REQUIRED = 'NEW USERNAME IS REQUIRED',
    USERNAME_CANNOT_BE_SAME = 'USERNAME CANNOT BE SAME',
    USERNAME_NOT_UPDATED = 'USERNAME NOT UPDATED',
    USERNAME_UPDATED_SUCCESSFULLY = 'USERNAME UPDATED SUCCESSFULLY',
    ID_IS_REQUIRED = 'ID IS REQUIRED',
    USER_ACCOUNT_DELETED = 'USER ACCOUNT DELETED'
}